Fixed many problems. Still having world gen issues...

+++ src/common/micdoodle8/mods/galacticraft/mars/GCMarsCaveGen.java
@@ -24,7 +24,7 @@ public class GCMarsCaveGen
 
     protected World worldObj;
 
-    public void generate(IChunkProvider par1IChunkProvider, World par2World, int par3, int par4, byte[] par5ArrayOfByte)
+    public void generate(IChunkProvider par1IChunkProvider, World par2World, int par3, int par4, int[] par5ArrayOfByte)
     {
         int var6 = this.range;
         this.worldObj = par2World;
@@ -44,7 +44,7 @@ public class GCMarsCaveGen
         }
     }
 
-    protected void recursiveGenerate(World par1World, int xChunkCoord, int zChunkCoord, int origXChunkCoord, int origZChunkCoord, byte[] par6ArrayOfByte)
+    protected void recursiveGenerate(World par1World, int xChunkCoord, int zChunkCoord, int origXChunkCoord, int origZChunkCoord, int[] par6ArrayOfByte)
     {
         int var7 = this.rand.nextInt(this.rand.nextInt(this.rand.nextInt(40) + 1) + 1);
 
@@ -70,12 +70,12 @@ public class GCMarsCaveGen
         }
     }
     
-    protected void generateLargeCaveNode(long par1, int origXChunkCoord, int origZChunkCoord, byte[] par5ArrayOfByte, double xPos, double yPos, double zPos)
+    protected void generateLargeCaveNode(long par1, int origXChunkCoord, int origZChunkCoord, int[] par5ArrayOfByte, double xPos, double yPos, double zPos)
     {
         this.generateCaveNode(par1, origXChunkCoord, origZChunkCoord, par5ArrayOfByte, xPos, yPos, zPos, 1.0F + this.rand.nextFloat() * 6.0F, 10.0F, 10.0F, -1, -1, 0.2D);
     }
 
-    protected void generateCaveNode(long par1, int origXChunkCoord, int origZChunkCoord, byte[] par5ArrayOfByte, double xPos, double yPos, double zPos, float par12, float par13, float par14, int par15, int par16, double heightMultiplier)
+    protected void generateCaveNode(long par1, int origXChunkCoord, int origZChunkCoord, int[] par5ArrayOfByte, double xPos, double yPos, double zPos, float par12, float par13, float par14, int par15, int par16, double heightMultiplier)
     {
         double var19 = (double)(origXChunkCoord * 16 + 8);
         double var21 = (double)(origZChunkCoord * 16 + 8);
@@ -227,7 +227,7 @@ public class GCMarsCaveGen
 
                                     if (var51 > -0.7D && var59 * var59 + var51 * var51 + var46 * var46 < 1.0D)
                                     {
-                                        byte var53 = par5ArrayOfByte[var48];
+                                        int var53 = par5ArrayOfByte[var48];
 
                                         if (var53 == (byte) GCMarsBlocks.marsGrass.blockID)
                                         {
